main: fix incorrect proxy detection in KDE when proxy is disabled globally
When the proxy is disabled globally in the System Settings in KDE the
proxy host and address settings may still be preserved in the kioslaverc/Proxy Settings/*Proxy
configuration settings.

Check the kioslaverc/Proxy Settings/httpProxy setting's value and
return None when it is not set to "1"(meaning that the "Use manually
specified proxy configuration" mode isn't selected).

Fixes #1452.

Signed-off-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>