branch: avoid adding branch details multiple times
Entering the name of a branch and then holding backspace can lead to a
situation where multiple BranchDetailsTask instances are running in
parallel.

This can lead to `_update_branches_finished()` getting called multiple
times, resulting in an entry with multiple text decorations.

Add a guard variable to ensure that only a single task is running at
any time.

Signed-off-by: David Aguilar <davvid@gmail.com>