branches: teach the filter field to actually filter refs
The filter field in the Branches tool was originally setup to make
matching branches and tags bold when the filter contains matching text.

Make the filter field filter the branches and tags instead. This makes
the behavior of the filter field match the behavior of the Status tool's
filter field and is generally a more intuitive behavior.

Closes: #1457
Suggested-by: @novebeta on github
Signed-off-by: David Aguilar <davvid@gmail.com>