stash: fix mutual exclusion of Stash options
The connect_checkbox() helper function was not correctly forwarding its
arguments, which was causing the callbacks to always receive None.

This could have been fixed by changing the lambda to call `get(widget)`
instead of the `get(checkbox)` typo, but simplifying the lambda so that
it direclty forwards its argument avoids having to re-query the value.

Related-to: #1463
Signed-off-by: David Aguilar <davvid@gmail.com>