commitmsg: move up and down visually when lines are wrapped
Adjust the keyPressEvent so that cursors are moved logically when word
wrapping is enabled. The keyPressEvent was previously skipping over the
first word-wrapped line when moving upwards and over the last line when
moving downwards.

Use QTextCursor::movePosition() to move the cursor logically and then
detect whether the cursor position changed before jumping to the
beginning or end of the line. The position won't change when we are at
the beginning and end, which allows us to narrow the scenarios under
which the jump behavior is triggered.

Signed-off-by: David Aguilar <davvid@gmail.com>