branch: always run the upstream branch actions
A defensive check to avoid doing redundant work prevented the
SetUpstreamBranch command from triggering. The checks can become invalid
when remote branches are renamed, so simply remove the check so that the
action is always triggered.

Additionally, the main button was not connected, which made it so that
only double-clicking the remote branch would trigger an action. Connect
the button to make the UI behave as expected.

Closes: #1475
Reported-by: Jacob Gustafson @Poikilos on github
Signed-off-by: David Aguilar <davvid@gmail.com>