dag: add an interface for selecting line ranges
Make it easy to select line ranges using the DAG viewer. Right-click on
a commit to open a Finder dialog where files and their contents can be
explored.

Select a file range using the contents window and use the right-click
action or the main dialog button to select the file range. The selected
file range is then added to the DAG viewer's revision text field.

Closes: #1483
Suggested-by: Michele Nalli <michelenalli96@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>