cmds: add support for chained commands
Extend the EditModel class so that a finalizer command can be configured
to run after the command is executed. This will allow us to chain
commands that would have run synchronously but are now running in the
background.

The CommandBus is used to transfer the command object from the
background thread back to the foreground thread before running it. This
is most useful for short-lived quick commands that modify the UI state.

Add chaining support to the Diff commands so that we can leverage
chaining when we switch to executing these commands in the background.

Related-to: #1484
Signed-off-by: David Aguilar <davvid@gmail.com>