cmds: prevent interleaved ContextCommands from clobbering state
Automatically ignore the results from long-running commands when a newer
command is run before that command completes. This will ensure that
commands that are run in the background end up updating state in the
correct order in the presence of long-running commands.

What this effectively does is makes it so that a long-running diff
command can be short-circuited by a later diff command that gets
launched after the slow diff.

This will be helpful when we switch to running diff commands in the
background to avoid blocking the UI.

Related-to: #1484
Signed-off-by: David Aguilar <davvid@gmail.com>