status: run diff commands in the background
Minimize the amount of time under which the main UI is blocked by
running diff commands in the background. The slow git commands will get
run in the background without holding onto the Python GIL.

This does not yet eliminate the slow speed of calling
QPlainTextEdit::setPlainText(...) and its blocking of the main UI when
called with very large text values, but it does eliminate the time spent
running git commands from contributing to this blocked UI time.

Related-to: #1484
Signed-off-by: David Aguilar <davvid@gmail.com>