prefs: do not specify --repo for the default history browser
Including the `--repo` argument can result in users storing this value
in their settings. This can lead to incorrect behavior when visualizing
branches.

Git DAG will inherit the current directory from the process that
launched it so there is no reason to include the `--repo` option.

A further enhancement would be to update how the history browser is
launched so that it supports a `${WORKTREE}` template variable that
users can use when configuring their settings. That would allow users to
specify the worktree location to external tools that do not
automatically load the repository from the current working directory.
That feature can be added in a future release.

Closes: #1496
Reported-by: Niel Buys <nbuys@ncomp.co.za>
Signed-off-by: David Aguilar <davvid@gmail.com>