dag: improve difftool behavior for merges
Detect the root commit in Git DAG's difftool interactions so that we can
use the {commit}^! syntax for just the root commit.

The {commit}~..{commit} syntax is now used for all other commits.

This ensures that the file list displayed by the files widget matches
the files that are displayed by difftool and improves the behavior when
diffing merge commits.

Closes: #1493
Reported-by: @pkzc on github
Signed-off-by: David Aguilar <davvid@gmail.com>