cmds: correct the diff text for images by deferring construction of finalizers
The model state fields need to be up to date when finalizers are
constructed otherwise they will snapshot the old state before the parent
command has been run. Not doing so caused the diff text display for
image files to not update correctly.

Defer construction of finalizer commands until after the parent command
has completed so that the finalizer command uses the updated model state.

Closes: #1498
Reported-by: Luke Horwell @lah7 on github
Signed-off-by: David Aguilar <davvid@gmail.com>