stash: improve keyboard interactions
Make the stash save input mode controlled using a combobox.
Add `Ctrl + {1,2,3}` hotkeys for switching the combobox modes.

Add `Ctrl + s` for saving stashes so that users can open the stash
dialog and hit this hotkey in order to save a stash. `Enter` also works
because the Save button is the default button.

Add a Help button and a `?` hotkey to display a help message that
includes the keyboard shortcuts.

Closes: #1506
Suggested-by: @birgersp on github
Helped-by: Luke Horwell @lah7 on github
Signed-off-by: David Aguilar <davvid@gmail.com>