models: use the correct message when re-merging a previously-aborted merge
When the commit message is explicilty set, e.g. when aborting a merge,
then we must clear the internal state of the commit messages that had
previously been seen.

Not clearing the state will prevent the (repeated) merge message from
being used because cola was incorrectly detecting this as a scenario
where the commit message was already previously applied.

Closes: #1519
Reported-by: Francesco Gallo <gallo.francesco.fra@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com=>