dag: improve performance in large repositories
Eliminate the use of locks as they are not actually needed in the
current implementation. These were needed back when a resume()
implementation existed and is no longer needed.

Increase the batch size from 512 commits to 1024 commits since empirical
testing on a larger repository (git.git) indicates that it improves the
load initial treewidget render time and overall load time since less
time is spent processing callbacks.

The final optimization is to defer the update of the graphview until all
of the commits have been gathered. Not doing so ends up locking up the
interface for much longer because the graphview is much slower to update
incrementally, unlike the treewidget which can incrementally update very
quickly.

The net result of these changes is that the startup time has improved
between 50-65% percent. The remaining slowness seems to be unavoidable
due to the use of Python (and the existence of the GIL).

Closes: #1526
Reported-by: @birgersp on github
Signed-off-by: David Aguilar <davvid@gmail.com>